Refactor VAD service into module and reorganize project
