+++ services/compress/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP compression service."""
