+++ services/vad/__init__.py
@@ -1,0 +1,1 @@
+"""WebSocket VAD service and utilities."""
